feat: Add estadoActivo parameter to FormCreateProductos
The code changes include adding the `estadoActivo` parameter to the `FormCreateProductos` component in the file `src/views/forms/FormCreateProductos.jsx`. This parameter is used in the `FormLayoutsWithIcon` component to render the form layout.

Based on the recent user commits and repository commits, the suggested commit message is "feat: Add estadoActivo parameter to FormCreateProductos".